requirements.txt
bs4
feedparser
lxml
psycopg2-binary
python-telegram-bot
requests
sqlalchemy
telethon